test: fix v1 usage in tests
